requirements.txt
streamlit==1.39.0
google-cloud-storage==2.18.2
pandas==2.2.3
plotly==5.24.1
regex==2024.9.11